Update after code review comments  - Fixes #74
