+++ src/main/groovy/deploydb/mappers/DeploymentMapper.groovy
@@ -5,8 +5,6 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import deploydb.WebhookModelMapper
 import deploydb.models.Artifact
 import deploydb.Status
-import deploydb.models.PromotionResult
-
 import javax.persistence.Entity
 
 /**
@@ -53,39 +51,6 @@ class DeploymentWebhookMapper implements WebhookModelMapper{
     }
 }
 
-class DeploymentPromotionWebhookMapper extends DeploymentWebhookMapper{
-
-    /**
-     *  Promotion result for webhook mapper
-     */
-    @JsonProperty
-    PromotionResult promotionResult
-
-    /**
-     * Default constructor
-     */
-    DeploymentPromotionWebhookMapper() { }
-    /**
-     *
-     * @param deployment Deployment to map for the webhook
-     * @param promotionResult Promotion result to map for the webhook
-     */
-    DeploymentPromotionWebhookMapper(deploydb.models.Deployment deployment,
-                            PromotionResult promotionResult){
-        this.id = deployment.id
-        this.artifact = deployment.artifact
-        this.service = deployment.serviceIdent
-        this.environment = deployment.environmentIdent
-        this.createdAt = deployment.createdAt
-        this.status = Status.VERIFIED
-        this.promotionResult = promotionResult
-    }
-
-    String toPayload() {
-        ObjectMapper mapper = new ObjectMapper();
-        return mapper.writeValueAsString(this) ;
-    }
-}
 
 
 /**
